// frontend/src/types.ts
export interface ProductInfo {
    serial_number: string;
    product_code: string;
    charge_point_id: string;
    ethernet_mac: string;
    bluetooth_mac: string;
    four_g_imei: string;
    master_card_rfid: string;
    slave_1_card_rfid: string;
    slave_2_card_rfid: string;
    product_description: string;
    bluetooth_key: string;
    bluetooth_iv_key: string;
    bluetooth_password: string;
} 